+++ HISTORY.rst
@@ -1,6 +1,8 @@
 3.1.11 (2019-04-XX)
 ===================
 * Fix multiple tag selection wizard step.
+* Change the required permission for the checkout info link from
+  document check in to document checkout details view.
 
 3.1.10 (2019-04-04)
 ===================